NRV 1773 Race V2

See merge request zan/odyssey!391

+++ src/odyssey/api/client/routes.py
@@ -37,7 +37,8 @@ from odyssey.api.client.models import (
     ClientWeightHistory,
     ClientWaistSizeHistory,
     ClientTransactionHistory,
-    ClientPushNotifications
+    ClientPushNotifications,
+    ClientRaceAndEthnicity
 )
 from odyssey.api.doctor.models import (
     MedicalFamilyHistory,
@@ -90,7 +91,9 @@ from odyssey.api.client.schemas import(
     SignAndDateSchema,
     SignedDocumentsSchema,
     ClientTransactionHistorySchema,
-    ClientSearchItemsSchema
+    ClientSearchItemsSchema,
+    ClientRaceAndEthnicitySchema,
+    ClientRaceAndEthnicityEditSchema
 )
 from odyssey.api.lookup.schemas import LookupDefaultHealthMetricsSchema
 from odyssey.api.staff.schemas import StaffRecentClientsSchema
@@ -114,25 +117,27 @@ class Client(Resource):
         #update staff recent clients information
         staff_user_id = token_auth.current_user()[0].user_id
 
-        #check if supplied client is already in staff recent clients
-        client_exists = StaffRecentClients.query.filter_by(user_id=staff_user_id).filter_by(client_user_id=user_id).one_or_none()
-        if client_exists:
-            #update timestamp
-            client_exists.timestamp = datetime.now()
-            db.session.add(client_exists)
-            db.session.commit()
-        else:
-            #enter new recent client information
-            recent_client_schema = StaffRecentClientsSchema().load({'user_id': staff_user_id, 'client_user_id': user_id})
-            db.session.add(recent_client_schema)
-            db.session.flush()
+        #if this request was made by a staff member, update their recent clients list
+        if staff_user_id != user_id:
+            #check if supplied client is already in staff recent clients
+            client_exists = StaffRecentClients.query.filter_by(user_id=staff_user_id).filter_by(client_user_id=user_id).one_or_none()
+            if client_exists:
+                #update timestamp
+                client_exists.timestamp = datetime.now()
+                db.session.add(client_exists)
+                db.session.commit()
+            else:
+                #enter new recent client information
+                recent_client_schema = StaffRecentClientsSchema().load({'user_id': staff_user_id, 'client_user_id': user_id})
+                db.session.add(recent_client_schema)
+                db.session.flush()
 
-            #check if staff member has more than 10 recent clients
-            staff_recent_searches = StaffRecentClients.query.filter_by(user_id=staff_user_id).order_by(StaffRecentClients.timestamp.asc()).all()
-            if len(staff_recent_searches) > 10:
-                #remove the oldest client in the list
-                db.session.delete(staff_recent_searches[0])
-            db.session.commit()
+                #check if staff member has more than 10 recent clients
+                staff_recent_searches = StaffRecentClients.query.filter_by(user_id=staff_user_id).order_by(StaffRecentClients.timestamp.asc()).all()
+                if len(staff_recent_searches) > 10:
+                    #remove the oldest client in the list
+                    db.session.delete(staff_recent_searches[0])
+                db.session.commit()
 
         #data must be refreshed because of db changes
         if client_data:
@@ -142,7 +147,12 @@ class Client(Resource):
        
         client_info_payload = client_data.__dict__
         client_info_payload["primary_goal"] = db.session.query(LookupGoals.goal_name).filter(client_data.primary_goal_id == LookupGoals.goal_id).one_or_none()
-        client_info_payload["race"] = db.session.query(LookupRaces.race_name).filter(client_data.race_id == LookupRaces.race_id).one_or_none()
+
+        race_info = db.session.query(ClientRaceAndEthnicity.is_client_mother, LookupRaces.race_id, LookupRaces.race_name) \
+            .join(LookupRaces, LookupRaces.race_id == ClientRaceAndEthnicity.race_id) \
+            .filter(ClientRaceAndEthnicity.user_id == user_id).all()
+
+        client_info_payload["race_information"] = race_info
 
         return {'client_info': client_info_payload, 'user_info': user_data}
 
@@ -182,7 +192,12 @@ class Client(Resource):
         # prepare client_info payload  
         client_info_payload = client_data.__dict__
         client_info_payload["primary_goal"] = db.session.query(LookupGoals.goal_name).filter(client_data.primary_goal_id == LookupGoals.goal_id).one_or_none()
-        client_info_payload["race"] = db.session.query(LookupRaces.race_name).filter(client_data.race_id == LookupRaces.race_id).one_or_none()
+        
+        race_info = ClientRaceAndEthnicity.query.filter_by(user_id=user_id).all()
+        for info in race_info:
+            info.race_name = LookupRaces.query.filter_by(race_id=info.race_id).one_or_none().race_name
+
+        client_info_payload["race_information"] = race_info
 
         return {'client_info': client_data, 'user_info': user_data}
 
@@ -1386,3 +1401,79 @@ class ClientWeightApi(Resource):
         health_metrics = LookupDefaultHealthMetrics.query.filter_by(age = age_category).filter_by(sex = sex).one_or_none()
         
         return health_metrics
+
+@ns.route('/race-and-ethnicity/<int:user_id>/')
+@ns.doc(params={'user_id': 'User ID number'})
+class ClientRaceAndEthnicityApi(Resource):
+    """
+    Endpoint for returning viewing and editing informations about a client's race and ethnicity
+    information.
+    """
+    @token_auth.login_required()
+    @responds(schema=ClientRaceAndEthnicitySchema(many=True), api=ns, status_code=200)
+    def get(self, user_id):
+        check_client_existence(user_id)
+
+        res = db.session.query(ClientRaceAndEthnicity.is_client_mother, LookupRaces.race_id, LookupRaces.race_name) \
+            .join(LookupRaces, LookupRaces.race_id == ClientRaceAndEthnicity.race_id) \
+            .filter(ClientRaceAndEthnicity.user_id == user_id).all()
+        
+        return res
+
+    @token_auth.login_required()
+    @accepts(schema=ClientRaceAndEthnicityEditSchema, api=ns)
+    @responds(schema=ClientRaceAndEthnicitySchema(many=True), api=ns, status_code=201)
+    def post(self, user_id):
+        check_client_existence(user_id)
+
+        if ClientRaceAndEthnicity.query.filter_by(user_id=user_id).first():
+            raise IllegalSetting(message=f"Race and ethnicity information for user_id {user_id} already exists. Please use PUT method")
+
+        for data in request.parsed_obj['data']:
+            if not LookupRaces.query.filter_by(race_id=data.race_id).one_or_none():
+                raise InputError(400, 'Invalid race_id.')
+
+            model = ClientRaceAndEthnicitySchema().load({
+                'race_id': data.race_id,
+                'is_client_mother': data.is_client_mother
+            })
+            model.user_id = user_id
+            db.session.add(model)
+        db.session.commit()
+        
+        res = db.session.query(ClientRaceAndEthnicity.is_client_mother, LookupRaces.race_id, LookupRaces.race_name) \
+            .join(LookupRaces, LookupRaces.race_id == ClientRaceAndEthnicity.race_id) \
+            .filter(ClientRaceAndEthnicity.user_id == user_id).all()
+
+        return res
+
+    @token_auth.login_required()
+    @accepts(schema=ClientRaceAndEthnicityEditSchema, api=ns)
+    @responds(schema=ClientRaceAndEthnicitySchema(many=True), api=ns, status_code=201)
+    def put(self, user_id):
+
+        check_client_existence(user_id)
+        
+        #remove existing race/ethnicity info
+        for data in ClientRaceAndEthnicity.query.filter_by(user_id=user_id).all():
+            db.session.delete(data)
+        db.session.commit()
+
+        #add incoming data from request
+        for data in request.parsed_obj['data']:
+            if not LookupRaces.query.filter_by(race_id=data.race_id).one_or_none():
+                raise InputError(400, 'Invalid race_id.')
+
+            model = ClientRaceAndEthnicitySchema().load({
+                'race_id': data.race_id,
+                'is_client_mother': data.is_client_mother
+            })
+            model.user_id = user_id
+            db.session.add(model)
+        db.session.commit()
+
+        res = db.session.query(ClientRaceAndEthnicity.is_client_mother, LookupRaces.race_id, LookupRaces.race_name) \
+            .join(LookupRaces, LookupRaces.race_id == ClientRaceAndEthnicity.race_id) \
+            .filter(ClientRaceAndEthnicity.user_id == user_id).all()
+
+        return res
